Add 'Marriage' article to sacraments section
A new article about the sacrament of marriage has been added to the sacraments section. This article includes details on the procedure and requirements for the sacrament, as advised by the church. The purpose of this addition is to provide clear and helpful information for parishioners seeking to understand or undertake the sacrament of marriage.

Signed-off-by: xyashino <szczypiropan123@gmial.com>