Add 'altar.mdx' with altar server's role and guidelines
This commit introduces a new file 'altar.mdx' under src/articles/groups. The file contains detailed information about the role of an altar server in liturgical ceremonies and standards for aspirants, servers, and lectors. The aim is to provide a comprehensive guide for those serving at the altar, covering their responsibilities, spiritual guidance, and principles for their behavior.

Signed-off-by: xyashino <szczypiropan123@gmial.com>