Add Anointing of the Sick article
A new article has been added to the sacrament section - Anointing of the Sick. This provides detailed information about the sacrament, its importance, when it's performed, and how one can avail it in our parish. Details also include the needed prep in cases where the sacrament needs to be administered to a sick person at their home. This is to ensure parishioners are informed and can get access to this sacrament when needed.

Signed-off-by: xyashino <szczypiropan123@gmial.com>