Remove image only if it has a path
Previously, the server would attempt to remove an image even if it didn't have a specified path. This could result in the server trying to delete a non-existent or wrong file, leading to unexpected exceptions and behavior. This change ensures that an image will only be removed if it has a valid path, greatly improving the reliability of the deletion process and reducing chances of potential errors.

Signed-off-by: xyashino <szczypiropan123@gmial.com>