Add 'value' props to Select component and SelectValue component
This commit introduces 'value' prop to Select component in selectStatusItem.tsx to ensure the correct value is displayed when the component is loaded. Also, in selectStatusTrigger.tsx, 'defaultValue' prop is added to SelectValue component to ensure that the initially selected value is displayed when the component loads. These changes enhance the usability of the SelectStatus component.

Signed-off-by: xyashino <szczypiropan123@gmial.com>