Refactor MdEditor component imports and props
In the MdEditor component, we've cleaned up the imports, removing an unused CSS import and rearranging the other lines for a more logical order. We've also made a minor adjustment to the Editor dynamic import, adding an extra comma but maintaining the ssr as false. In addition, the 'theme' prop was removed from the Editor component as it was unnecessary and unused.

Signed-off-by: xyashino <szczypiropan123@gmial.com>