Replace AdminPageTitle with PageTitleWithPrevBtn in admin components
The AdminPageTitle component was replaced with the more versatile PageTitleWithPrevBtn component across all admin pages for improved user navigability. This new component not only displays the page title but also includes a previous button, enabling users to effortlessly navigate back to the previous page. This change enhances the application's usability and user experience.

Signed-off-by: xyashino <szczypiropan123@gmial.com>