Expand data returned in week intentions functions
Added additional fields to the selected data in the functions getWeekIntentionsWithRelations and getActiveWeekIntentions. The changes were made to improve the data returned when calling these functions, providing a more complete set of information about the week and day. This modification also ensures consistency in the data being fetched across different parts of the application. Specifically, in the getActiveWeekIntentions function, instead of directly fetching week intentions, it now calls the getWeekIntentionWithRelations function that returns more comprehensive data.

Signed-off-by: xyashino <szczypiropan123@gmial.com>