Refactor week-intentions-store interface definitions
This commit focuses on refactoring type definitions in the week-intentions-store. The revisions improve the reusability of the code by replacing hard-coded object interfaces with TypeScript's utility type 'Omit'. This allows for type inference of Day and Intention entities, leading to a cleaner and more maintainable codebase. Also, the file extension was changed to '.d.ts' to treat it as a TypeScript declaration file for better modules segregation and code transparency.

Signed-off-by: xyashino <szczypiropan123@gmial.com>