Add utility to convert API response to store data
Introduced a new utility function 'convertIntentionsResponseToStoreData' in 'src/lib/utils'. This function translates the API response related to 'WeekIntentions' into a format suitable for updating the application store. By creating this function, data from the back-end can be manipulated into a structure that optimises rendering in the frontend, improving UX.

Signed-off-by: xyashino <szczypiropan123@gmial.com>