Rename and update week-intentions-store interface
The file 'week-intentions-store.interface.d.ts' has been renamed to 'week-intentions-store.d.ts' and updated with added type definitions. The changes were made to enhance the type handling for the 'dayId' and 'weekId' fields in the 'OneIntention' and 'DayIntentions' types, respectively, allowing them to accept null values. A 'Weekday' field was also added to the 'DayIntentions' type for clearer definitions.

Signed-off-by: xyashino <szczypiropan123@gmial.com>