Fix typecasting in get-week-intentions-with-relations
The function `getWeekIntentionsWithRelations` was changed to explicitly typecast its return value as `WeekIntentionsWithRelationsResponse`.

The purpose of typecasting here is to enforce data integrity and help eliminate possible type-related runtime errors. This change was necessary because the response from the `findUnique` function of the `client.weekIntentions` object might not always perfectly match the `WeekIntentionsWithRelationsResponse` type structure. Thus, we ensure the return type conformity.

Signed-off-by: xyashino <szczypiropan123@gmial.com>