Refactor common utilities to use 'WeekIntentionsDay'
The module common.ts was adjusted to use the WeekIntentionsDay type instead of the previous Day type from "@prisma/client". Also, an unnecessary 'Day' import was removed. This change was necessary to fit the new structure of weekly schedulings, allowing us to interact with the DB entries in a more coherent manner.

Signed-off-by: xyashino <szczypiropan123@gmial.com>