Refine Navbar logo's focus state styles for visibility
Made CSS changes to the admin navigation bar logo in AdminNavbarHeader.tsx, adding focus state styles to enhance accessibility. This change will facilitate distinguishability of the logo when it is focused, making it more accessible for keyboard-only users. The upgrade complies with the standards and best practices of accessible web design.

Signed-off-by: xyashino <szczypiropan123@gmial.com>