Update Logo component styling in login page
Modified the className props of the Logo component in the LeftSection of the login page. These changes were made to fix the hover effect on the logo. Previously, it was changing the background color on hover. With the applied changes, instead of the background color, the text color changes on hover, enhancing user experience.

Signed-off-by: xyashino <szczypiropan123@gmial.com>