Refactor PageTitleWithPrevBtn component's prop interface
Interface 'Props' for component 'PageTitleWithPrevBtn' has been renamed to 'PageTitleWithPrevBtnProps' for better clarity. The interface is now also exported to be reused in other components if necessary. The destructured prop arguments in the component function also reflect this change. This ensures to maintain consistency in naming across the application and improve readability and understanding of codebase.

Signed-off-by: xyashino <szczypiropan123@gmial.com>