Refactor weekIntentionsTabContent props interface
The weekIntentionsTabContent component props interface previously consisted of the 'Day' and 'Intention' types imported from Prisma client. This led to a lot of unnecessary code and complexity in managing types across several files. In this commit, these components are replaced by a 'WeekIntentionsDay' interface that includes all the necessary attributes intended for the component. This refactor improves maintainability and understanding of the component props structure.

Signed-off-by: xyashino <szczypiropan123@gmial.com>