Refactor week-intentions constants and type imports
This commit refactors the 'week-intentions' constants, particularly the `dayIntentions` Map object. The 'intentions' arrays in this object have been typed explicitly as `OneIntention[]`. This choice of explicit typing helps to improve code readability and maintainability.

Also, the import statement at the top of the file has been altered to import `OneIntention` from '@/types/week-intentions-store' which provides a clear indication of where these types are being pulled in from.

Signed-off-by: xyashino <szczypiropan123@gmial.com>