Fix tab selection and key assignment in ModifyWeekIntentions
The defaultValue property was replaced with value to synchronize the active tab with the activeDay state. This ensures that the correct tab is active when the component renders. Moreover, the id extraction method was adjusted in ModifyDayIntentions' key prop by prioritizing value.id over id, to guarantee access to unique intent id even when an id is not explicitly provided.

Signed-off-by: xyashino <szczypiropan123@gmial.com>