Add error handling for album creation
This commit adds a check for the result of the `createAlbum` function in the Albums API route. If the result is falsy (e.g. `null` or `undefined`), it indicates that the album creation has failed, so we now return an error response. This additional check helps to prevent further operations like the `createDirectory` function call, which is dependent on successful album creation, thus improving the reliability of the API.

Signed-off-by: xyashino <szczypiropan123@gmial.com>