Refactor fetching method for active announcements
This commit replaces the older method for fetching active announcements with a newer, more efficient function. The function 'getActiveAnnouncement' was removed and replaced with an 'apiCall' function to fetch Announcement details. This is in line with our goal of enhancing our code by making API calls more standardized and flexible. Also, added a handling case for situations where no announcement is available.

Signed-off-by: xyashino <szczypiropan123@gmial.com>