Update validation messages in schemas
The validation messages for empty fields in the 'categories' and 'album' schemas have been updated. They are now more descriptive, informing users that the 'name' in the category schema and 'title' in the album schema are required fields. This change aims to enhance user experience by giving them clear error messages.

Signed-off-by: xyashino <szczypiropan123@gmial.com>