Add AlbumImage and AlbumImageTrigger components
These new components were added to manage the visualization of images within an album. The AlbumImage component takes care of the image's display and loading state. On the other hand, the AlbumImageTrigger component toggles visibility of the image to fullscreen on user interaction. These components improving the user experience by making it easier to view and navigate through the album images.

Signed-off-by: xyashino <szczypiropan123@gmial.com>