Update AlbumImage to maintain aspect ratio
This commit modifies the AlbumImage component to maintain the aspect ratio for images. Previously, the images were set to take up the full width and height of its container. The `aspect-video` class has now been added to maintain the aspect ratio of the images, providing a better visual experience for users.

Signed-off-by: xyashino <szczypiropan123@gmial.com>