Add revalidation on API and update admin view
This commit adds revalidation of the Tag INTENTIONS in different API endpoints by importing and using the revalidateTag function from the next/cache library. Each time a GET, UPDATE or DELETE operation is performed on the intentions, the tag is revalidated, ensuring that stale data is not served.

In the admin view, the title of the WeekIntentions component has been updated to include the intention's uuid, allowing for more specific context when editing. The existing weekIntention object is also logged to console for debugging purposes.

Signed-off-by: xyashino <szczypiropan123@gmial.com>