Remove unused database methods
This commit removes all the unused database methods in the lib/db directory. These functions were no longer necessary due to the restructuring of the application's data handling strategy. The change is needed to keep the code clean and maintainable, it also reduces the complexity of the App by removing unnecessary code.

Signed-off-by: xyashino <szczypiropan123@gmial.com>