Add Week Intentions schema and related tables
A new file, src/db/schema/week-intentions.ts, has been created to define the schema for WeekIntentions, Day, and Intention tables in Postgres database. The file also contains the necessary relationships between these tables. This is to setup a DB structure for tracking users' weekly intentions and the status of each day.

Signed-off-by: xyashino <szczypiropan123@gmial.com>