Add constants for database enums
This commit introduces a new file that contains two enums, `STATUS` and `DAY`, for use within the database. It aims to provide uniformity in data entry, promote code reuse, and simplify the retrieval and manipulation of data. The `STATUS` enum will represent different states of an entity while `DAY` enum will be used to represent the week's days.

Signed-off-by: xyashino <szczypiropan123@gmial.com>