Replace Weekday import with DAY constant in useWeekIntentionsStore
In the useWeekIntentionsStore.ts file, the import from prisma client `Weekday` was replaced with the `DAY` constant from the local constants directory. This change will ensure that the days of the week are no longer tied to the Prisma client, providing greater flexibility in week date handling and contributing to more effective decoupling of components.

Signed-off-by: xyashino <szczypiropan123@gmial.com>