Add server and client environment configuration
Implemented a new environment configuration for both server and client in two separate TS files. This change was necessary to validate and parse environment variables using `Zod` and separate the concerns of server and client configurations. The configuration ensures the presence and validity of necessary environment variables. In addition, a preprocessing function was used for the ALLOW_EMAIL_LIST variable to parse comma-separated emails from a string.

Signed-off-by: xyashino <szczypiropan123@gmial.com>