"Refactor FetchDataWithToast to improve error handling
The FetchDataWithToast function is enhanced to get better readability and error handling. The code now includes notFound rendering for 404 responses when not in EdgeRuntime, and also utilizes the env variable instead of directly using process.env for API URLs. Error messages are also thrown with potential parsed server error responses, instead of the rudimentary 'Something went wrong!' message. Improved handling of result types with expecting a Record type for the generic 'T' in apiCallWithToast."

Signed-off-by: xyashino <szczypiropan123@gmial.com>