Refactor import statements in useAlbumCoverImageStore
The import statements for getImageCoverPath and updateCoverImage functions have been compacted into a single line. The name of the file from which these functions are imported has also been shortened from "api-calls" to "api" to make it more concise and understandable. This change improves readability and understanding of the code, making it easier to maintain.

Signed-off-by: xyashino <szczypiropan123@gmial.com>