Add new handlers for week-intentions, album, administrator & announcement
In this commit, several new handlers were added in the `src/db/handlers` directory for different modules of our application. This was done to structure the code better and make it maintainable. The new files include handlers for `week-intentions`, `album`, `administrator` and `announcement` modules. Each file contains related queries and operations that would be performed on their respective tables in the database.

Signed-off-by: xyashino <szczypiropan123@gmial.com>