Refactor services structure and simplify usage
This commit refactors the structure of the services by removing the unnecessary "crud" subfolders and renaming the service files to directly represent the related resource. This change simplified the project structure and made the services easier to use. Furthermore, for better encapsulation and consistency, related API or server method calls were grouped together under the appropriate service. Environment variables were also migrated to a common configuration file to enhance maintainability.

Signed-off-by: xyashino <szczypiropan123@gmial.com>