Remove announcements crud service
The 'announcements/crud.ts' file was deleted. This file contained a service for CRUD operations for announcements. The decision to remove it is based on the redesign of the features related to announcements, where this kind of CRUD service is no longer necessary. Also, its removal contributes to simplifying the overall project structure.

Signed-off-by: xyashino <szczypiropan123@gmial.com>