Update import statements and refactor API call in albums page
Modified the import statements in the albums page to match the latest updates in the types of the db in `@/types`. Removed the unnecessary import of `notFound` from `next/navigation`.

Refactored the API call in `generateStaticParams` function to use `AlbumListResponse` instead of `AlbumsResponse`. Also code is updated to make sure that `images` array is treated as `ImageListResponse` for mapping operation. Redundant condition for returning `notFound` is also removed. These changes are made keeping in line with the recent updates in types in `@/types/db/album`.

Signed-off-by: xyashino <szczypiropan123@gmial.com>