Handle null value for intention and status in week intentions
This commit adds handling for null values in week intentions. Previously, if "intentionData.value" or "weekIntentions.status" was null, the program could crash or behave unexpectedly. Now, "?" operator is used to ensure that if the value is null, it is replaced with an empty string in case of "intentionData.value" and "NONE" for "weekIntentions.status", ensuring stable behaviour. Removed unnecessary import of "Status" from "@prisma/client".

Signed-off-by: xyashino <szczypiropan123@gmial.com>