Refactor API calls in admin gallery page
API calls in the admin gallery page were refactored for more clarity and efficiency. The `Category` import was removed as it was found to be unnecessary. Also, broadened the specificity of the expected response type from API calls, by replacing `AlbumsResponse` and `Category` with `AlbumResponse[]` and `CategoriesResponse`. This change is meant to reflect the actual structure of the response data received. Also, a log statement was added for debugging purposes.

Signed-off-by: xyashino <szczypiropan123@gmial.com>