Refactor week intentions API call
The API call for fetching weekly intentions in the admin section was refactored. Instead of using the `getManyWeekIntentions` method, the more flexible `apiCall` utility is being used directly. This change provides additional context to the API call, such as the use of the `RevalidateTag.INTENTION` which can significantly improve the efficiency of cache revalidation.

Signed-off-by: xyashino <szczypiropan123@gmial.com>