Refactor import and update schema types
Updated the import syntax to import multiple items on a single line, making it more concise and easier to read. Also, the type of 'weekId' and 'dayId' was changed from uuid to varchar in the 'week-intentions' schema to align with the actual type in the associated tables. These changes improve overall code readability and fix the type mismatching issue in the schema.

Signed-off-by: xyashino <szczypiropan123@gmial.com>