Refactor API response format and adjust variable names
This commit replaces all instances of NextResponse.json with new Response that includes JSON.stringify. This change ensures that strings are correctly formatted as JSON in responses. Also, the local scoped variable 'album' in the DELETE request handler of administrators' route was incorrectly named. The variable retrieves an administrator object, and has thus been renamed to 'administrator'.

Signed-off-by: xyashino <szczypiropan123@gmial.com>