Refactor client to use Db Handlers instead of API calls
The client of both 'intentions' and 'announcements' pages was refactored to directly interact with the database handlers rather than using API calls. This change simplifies the code and eliminates the need for unnecessary dependencies on API routes and revalidation tags. Additionally, for 'announcements', if there are no active announcements a 'notFound()' is now returned to handle empty results gracefully.

Signed-off-by: xyashino <szczypiropan123@gmial.com>