Simplify path revalidation and remove admin path references
Refactored the path revalidation procedure used in the routes of different sections of the application. Removed revalidation for admin paths which was unnecessary and causing inefficient revalidations. The change is part of our continual optimizations of the code base, to streamline processes and improve the application's overall performance. The revalidations are now focused on the Client paths or public facing sections improving our resource management, reducing wasteful operations.

Signed-off-by: xyashino <szczypiropan123@gmial.com>