Update image source paths in patron.mdx and LeftSection.tsx
The paths to the 'kolbe' and 'auth' images in the patron.mdx file and LeftSection.tsx component respectively have been altered. This was done to bring consistency to the code base as all static files including images are now being housed in a single 'images' folder. Scaling up, this will enhance easier management and retrieval of static assets.

Signed-off-by: xyashino <szczypiropan123@gmial.com>