Add unauthorized page
A new page for displaying the 401 Unauthorized error was added. This serves to inform the user when they do not have access permissions to a page. The page includes a message of the error and two buttons, one to retry access and another to go back to the home page. This change improves user experience by providing relevant feedback and options when an unauthorized access occurs.

Signed-off-by: xyashino <szczypiropan123@gmial.com>