Refactor environment variable usage in ProviderButton
Instead of getting the NEXT_PUBLIC_APP_URL from process.env, refactored the code to get it from the 'env' module. This makes the codebase consistent with the rest of the application where environment variables are managed by the 'env' module. This change also removed unnecessary ternary operator, making the code cleaner and more readable.

Signed-off-by: xyashino <szczypiropan123@gmial.com>